Split up more frontend scripts
