+++ website/src/BaseLayout.astro
@@ -117,5 +117,8 @@ const { default: favicon } = await import("/public/img/favicon.svg?raw");
 		<script>
 			import "./frontend-scripts/index.js";
 		</script>
+		<script>
+			import "./frontend-scripts/docsearch.js";
+		</script>
 	</body>
 </html>
