+++ website/src/frontend-scripts/toc.ts
@@ -1,31 +1,9 @@
-import "./docsearch";
-import {matchesDark, getCurrentTheme} from "./util";
-
-//# Responsive width
-let isMobile = false;
-window.addEventListener("DOMContentLoaded", () => {
-	const mobileMatchMedia = matchMedia("(max-width: 768px)");
-	isMobile = mobileMatchMedia.matches;
-
-	mobileMatchMedia.addListener((e) => {
-		isMobile = e.matches;
-
-		// Close the mobile sidebar when switching from mobile to desktop
-		if (isMobileNavVisible && !isMobile && isMobileNavVisible) {
-			toggleMobileSidebar();
-		}
-	});
-});
-
-//# Table of Contents
 const originalTitle = document.title;
 const headerMobile = document.querySelector(".header-mobile");
 
-/** @type {HTMLElement}*/
-const tocSidebar = document.querySelector(".toc-sidebar");
+const tocSidebar: HTMLElement | null = document.querySelector(".toc-sidebar");
 
-/** @type {Array<HTMLAnchorElement>}*/
-const tocLinks = Array.from(document.querySelectorAll(".toc a"));
+const tocLinks: HTMLAnchorElement[] = Array.from(document.querySelectorAll(".toc a"));
 
 /** @type {Array<{
  * heading: HTMLElement,
@@ -431,127 +409,5 @@ if (tocLinks.length > 0) {
 
 	window.addEventListener("DOMContentLoaded", () => {
 		manager.attach();
-	});
-}
-
-//# Team list shuffle
-
-/**
- * @template T
- * @param {Array<T>} array
- * @returns {Array<T>}
- */
-function randomShuffle(array) {
-	let count = array.length;
-	let temp;
-	let index;
-	while (count) {
-		index = Math.floor(Math.random() * count--);
-		temp = array[count];
-		array[count] = array[index];
-		array[index] = temp;
-	}
-	return array;
-}
-
-const creditsPeopleLists = document.querySelectorAll(".credits-people-list");
-for (const list of creditsPeopleLists) {
-	const items = list.querySelectorAll("li");
-	for (const li of randomShuffle(Array.from(items))) {
-		list.appendChild(li);
-	}
-}
-
-//# Code expanders
-
-const collapsed = document.querySelectorAll("pre.collapsed");
-for (const elem of collapsed) {
-	elem.addEventListener("click", () => {
-		elem.classList.remove("collapsed");
-	});
-}
-
-//# Color scheme switcher
-
-function toggleColorSchemeSwitch(evt) {
-	const currentScheme = getCurrentTheme();
-	const newScheme = currentScheme === "dark" ? "light" : "dark";
-	window.localStorage.setItem("data-theme", newScheme);
-	evt.currentTarget.setAttribute("aria-checked", newScheme === "dark");
-	document.documentElement.classList.add("transition");
-	document.documentElement.setAttribute("data-theme", newScheme);
-	onColorSchemeChange();
-}
-
-function onColorSchemeChange() {
-	window.dispatchEvent(new Event("colorschemechange"));
-}
-
-const colorSchemeSwitcher = document.querySelector(".color-scheme-switch");
-// rome-ignore lint/js/preferOptionalChaining: netlify's node version does not support optional call expressions
-if (colorSchemeSwitcher != null) {
-	colorSchemeSwitcher.addEventListener("click", toggleColorSchemeSwitch, false);
-}
-
-matchesDark().addEventListener("change", () => {
-	onColorSchemeChange();
-});
-
-//# Mobile navigation
-
-const mobileSidebarHandle = document.querySelector(".mobile-handle");
-const mobileActiveTargets = document.querySelectorAll(
-	".page-header, .page-header-mobile, .docs-sidebar",
-);
-let isMobileNavVisible = false;
-function toggleMobileSidebar() {
-	isMobileNavVisible = !isMobileNavVisible;
-	mobileSidebarHandle.classList.toggle("active");
-	document.body.classList.toggle("no-scroll");
-	if (isMobileNavVisible) {
-		for (const elem of mobileActiveTargets) {
-			elem.classList.add("mobile-active");
-		}
-	} else {
-		for (const elem of mobileActiveTargets) {
-			elem.classList.remove("mobile-active");
-		}
-	}
-}
-// rome-ignore lint/js/preferOptionalChaining: netlify's node version does not support optional call expressions
-if (mobileSidebarHandle != null) {
-	mobileSidebarHandle.addEventListener(
-		"click",
-		(event) => {
-			event.preventDefault();
-			toggleMobileSidebar();
-		},
-		false,
-	);
-}
-
-//# Header scrolls to top
-let topAnchors = Array.from(document.querySelectorAll("[href='#top']"));
-if (location.pathname === "/") {
-	topAnchors = [...topAnchors, ...document.querySelectorAll(".logo")];
-}
-for (const elem of topAnchors) {
-	elem.addEventListener("click", (e) => {
-		if (window.scrollY > 0) {
-			e.preventDefault();
-
-			if (tocSidebar != null) {
-				tocSidebar.scrollTop = 0;
-			}
-
-			window.scrollTo(0, 0);
-
-			// Remove the hash
-			history.pushState(
-				"",
-				document.title,
-				window.location.pathname + window.location.search,
-			);
-		}
 	});
 }
